backend/requirements.txt
pandas
fastapi[standard]
uvicorn[standard]
python-dotenv
python-multipart
google-adk
pyarrow
python-dateutil
s3fs
PyJWT
postgres-mcp
psycopg2-binary
